Created before_filter to make a @users instance variable available to creation/edition actions
I originally created an instance variable for :new only and realized that since the validation actually happens on :create the @users disappeared after validation.

It's repeated between expenses & tasks though (the redundancy, it hurts!) so unless we find a way to extend expenses from a basic task model we'll just have to put the @users list in the application helper. Might come handy...