Restyled tables.
Removed borders (box-shadows are better and more modular) and text-indents (should never be used on a table as they don't account for text wrapping the the next line).

Inverted the background gradient on header to distinguish them from normal data rows, and used color variables with SassScript for all blue tones.